No warnings

+++ podster/SVSubscriptionGridViewController.m
@@ -166,7 +166,7 @@ static const int ddLogLevel = LOG_LEVEL_INFO;
         NSError *error;
         NSArray *newItems = [self.context executeFetchRequest:request error:&error];
         NSAssert(error == nil, @"There was an error while fetching the next unplayed item:%@", error);
-        DDLogVerbose(@"Retrieved %lu items for display", newItems.count);
+        DDLogVerbose(@"Retrieved %lu items for display", (long)newItems.count);
         self.noContentLabel.text = NSLocalizedString(@"FAVORITES_NO_CONTENT", @"Message to show when the user hasn't added any favorites yet");
         self.noContentLabel.numberOfLines = 0;
